fixed typo and close the project

+++ public/application/controllers/DeleteItem.php
@@ -8,12 +8,21 @@ class DeleteItem extends CI_Controller
         parent::__construct();
         $this->load->database();
         $this->load->model('components/Delete');
+        session_start();
     }
 
     public function index()
     {
-        $ItemSN = $this->input->get('ItemSN');
-        $this->Delete->deleteItemWhereItemSN($ItemSN);
-        header('Location: ItemList');
+        if (isset($_SESSION) && $_SESSION['logged_in'] == TRUE) {
+            if ($_SESSION['access'] == 0) {
+                $ItemSN = $this->input->get('ItemSN');
+                $this->Delete->deleteItemWhereItemSN($ItemSN);
+                header('Location: ItemList');
+            } else {
+                header('Location: Home');
+            }
+        } else {
+            header('Location: Login');
+        }
     }
 }
